+++ Magic/CHANGELOG.md
@@ -9,6 +9,7 @@
  - Add `has_decimals` builtin_currency option to override currency formatting
  - Add `custom_formatting` option to builtin_currency.currency to turn off using Vault formatting (which is now again the default)
  - Fix an issue where hitting a magically-spawned falling block with another spell before it lands could make it stay permanently
+ - Add `model` config option to mobs, for ModelEngine support
 
 # 8.6.2